+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/viewmodel/channel/ChannelInfoViewModel.kt
@@ -53,7 +53,7 @@ public class ChannelInfoViewModel(
     },
 ) : ViewModel() {
 
-    private val controller: ChannelInfoViewController get() = controllerProvider()
+    private val controller: ChannelInfoViewController by lazy { controllerProvider() }
 
     /**
      * @see [ChannelInfoViewController.state]